chore: optimize checks
- [x] Add health checks for PR
- [x] Remove windows and macos for code quality checks
- [x] Trigger deploy on demo change
- [x] Run e2e only on main push

+++ packages/e2e/fixtures/index.ts
@@ -56,11 +56,15 @@ export const test = base.extend<TestExtension>({
       });
 
       await use(context);
-
       await deleteCryptKeeperStorage(context);
-      await context.close();
+
+      if (!fs.existsSync(downloadsPath)) {
+        return;
+      }
 
       fs.rmSync(downloadsPath, { recursive: true, force: true });
+
+      await context.close();
     },
     { scope: "test" },
   ],
